requirements.txt
argcomplete
attrs
numpy
pytest
neo4j
transformers

